feat: agregados tests y fixes generales

+++ TpGrupo12OO2/src/dao/PedidoDao.java
@@ -109,7 +109,7 @@ public class PedidoDao {
 	}
 
 	// Traer Pedido por id
-	public Pedido traer(int idPedido) {
+	public Pedido traer(long idPedido) {
 
 		Pedido objeto = null;
